Continued work on lazy loading
Currently chunks are stored in state after initial fetch.
These chunks are sent to the view (cardgrid) based on an index,
this index is also set in state.

Attempts to update the index using events fired on scroll
are unsuccessful, this is because the scroll event fires
far too many times when it meets the points set by our
function.

WIP