Continuing work on lazy loading
We are building out the functionality of the lazy
loading. A bug has been discovered where we
had accidentally assigned a function to be called on
each card in the grid, this resulted in the function
being called 16* times.

Functions and functionality moved out to the CardGridComponent,
additionally the CardGridComponent has been turned into a
class, stateful component.